[Minor][Testing] Consolidate IRs into corresponding functions
We moved most of the IR definition into the testing methods correspondingly.
